Refactor test suite
Refactor test suite
