+++ package.json
@@ -49,7 +49,6 @@
         "transform": {
             "^.+\\.(t|j)sx?$": "ts-jest"
         },
-        "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
         "moduleFileExtensions": [
             "ts",
             "tsx",
